- refactor run - add travis - add init test
