+++ .travis.yml
@@ -1,0 +1,16 @@
+language: python
+cache:
+  directories:
+    - "${HOME}/virtualenv"
+
+python:
+  - "3.3"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+
+install:
+  - pip install -r requirements.txt
+
+script:
+  - pytest --cov=app/
