+++ README.MD
@@ -1,6 +1,8 @@
 # Web client for manage ElasticMQ
 
-###Getting started
+### Getting started
+[![Requirements Status](https://requires.io/github/ArtyomAnohin/ElasticmqUI/requirements.svg?branch=feature%2Fadd_error_alert)](https://requires.io/github/ArtyomAnohin/ElasticmqUI/requirements/?branch=feature%2Fadd_error_alert)
+[![Build Status](https://travis-ci.org/ArtyomAnohin/ElasticmqUI.svg?branch=master)](https://travis-ci.org/ArtyomAnohin/ElasticmqUI)
 
 Install docker and run:
 ```
