+++ start.py
@@ -1,4 +1,5 @@
-from app import app
+from app import create_app
 
 if __name__ == "__main__":
-    app.run(debug=True, host='0.0.0.0')
+    app = create_app(debug=True)
+    app.run(host='0.0.0.0', port=5000)
